admin: fix problem to delete account
There was a problem with deleting an account that once have a detailed
schedule: its deletion was included at account deletion, but its
details deletion wasn't.

Add tests for it and for trying to delete account with ACTIVE
schedules (it should not be possible, one should disable the schedule
first, to avoid misunderstoods).