money: fix details update
Before, when you tried to update a move that originally had simple value
adding both details and simple value, the nhibernate used to crash.

Now the system gives the proper answer, plus keep moves from being
created with both data.

On MVC, the interface should handle this, so the proper handling was
implemented.