tests: add check of move description - get/report
A schedule can be marked to show installments on its children moves.

When you get the month report, moves that belong to this kind of
schedule should have this number on its description.

But when you want to get a move to edit, this is not editable part of
description, so it should not come within description field.