dev-helper: fix ip setting
For mobile:

To make it easier to develop, the site changes the ip pointed by app
when it starts. The old logic controls when to do it by checking the
environment. Now it is checked by if the file exists at the path only.

Given the IP class was called only by Mobile file, that just called it,
the files were unified.

For website:

The setting need to be changed at applicationhost.config. Couldn't find
a way to configure this at the project, so I made the project change it
at startup, at the same file that changes the android ip to communicate
with the local site.