android: move files back to lots-of-folders structure
IDE does not understand well the struct if you use a folder with dots.
When you use shortcuts to create files, it creates de files in the
wrong folder.

Dreaming about Android Studio understanding more concise structure and
knowing that it's kotlin. At tests, not even the external folder can be
called kotlin.