email: fix error with session
When the error happens, the result of the error report e-mail sending
will be put inside the object that identifies the user in the
communication with the browser. It is done to access the result at the
error page and show a proper message, saying whether the support
received an e-mail or not.