tests: fix error checking
It was verifying if the expected error, the enum that comes from the
text, was not null.

1) It is an enum, it will never be null
2) It should check the recorded error

There was no problem in test results because the next line would throw
an exception if the error is null, but this was not the right way to
show the problem with the tests.