robot: remove not needed logic
That was using Save method to save schedule after increasing deleted
moves count. This method checks if the Schedule being saved is
respecting the Categories rule defined by the user.

In a case that an user add an Schedule while using categories, stop
using categories and try to delete a Move from the Schedule, it would
result an error, because it is trying to edit a Schedule with Category.

But this check is not necessary here, because it is just updating the
deleted count field. So, removed all the overkill.

Given UpdateConfigInternal is not used anywhere else but inside ONE
method AdminService, put its code inside the method.