android: replace close button by contact button
The close button is not needed, given android way of closing is usually
click at the button that lists the apps and then close the ones you want
to close.

This was added given a collegue of mine saying they always hate when an
app does not have a close button, like 7 years ago.

No android app has a close button.

But the contact button is something is missed. If you have a problem,
you want a fast way to contact the people that can solve it.