docs: remove task that could bring misleading data
If you delete a move and just remove the line, the total value of the
account will be outdated. If you update the value locally, something
could be changed at the site and the moves could be outdated.

If you need to bring everything again, it is better to refresh the
entire page.

Thanks, @konkah