docs: recalculate dif/imp/pts
If a task depend on another, it's difficult cannot be lower, because the
difficult includes making the other. Same time, if a task have a
dependent one that is more important, so the importance of it increases,
given it is needed to the other one to be done.

In this case, the importance of tip was rethought, given the importance
of other tasks that are above it.