publish: change task description
I will not use ECS for now: it became too hard and expensive to test
stuff to upload at it, and AWS does not allow to cancel the reserved
instance. The only thing they accepted is changing it to linux, which
costs 60% of the windows one.

Given I have this instance for 3 years now, I will leave to use ECS when
I have more time and money for this.