android: send moves after internet is back
If you are offline, the app keeps the move data and start a server to
send the moves to the site when you are online again.

In case your move have a problem, it shows the move to you to change and
try to save again next time you enter at the moves screen.