robot: fix disabling schedule when closing account
The schedule disabling was being made by changing the property and
calling the method that saves the schedule the first time.

This method was checking Value and Detail. This information is stored
duplicated at DB to make it easier to recover. The Value is the total of
the details.

But this information should not be given duplicated by those who try to
save the move or the schedule.

The external disabling schedule way was fixed, not needing the checks
that is used when you are really creating the schedule. The disable by
closing account now is using same method.