legal: set initial lastAccess value to deploy time
Everyone who is using the system will be checked once the robot is
online. If some user is already more than 90 days without accessing:
- the first day after deploy they will receive a 30 days warning;
- the second day they will receive a 60 days warning;
- and the third day they will be deleted.

To give the same time to old users to check and decide as the time new
users will have, the system must set deploy time as the last time the
person interacted with the system.