site: show activate warning only after 3 days
If the warning is already there the first time the person logs in, it
can cause the impression that this is just part of the layout of the
site. At the same time, the first time the person is seeing the system
they don't need to imediately activate by e-mail, so there is no reason
to bother the person with this.

The last 2 days the warning is more showy.