chore: replace magic numbers
The magic number concept is putting a number inside your code without
explaining for what. The numbers get lost inside the code. Putting a
variable for the number at the beginging helps to identify what is the
number and acts like configuration, so you don't have to find the code
if you need to adjust the number.