tests: move a generic test suit from safe folder
The tests should be moved to the place when its action happens. For
example, the first test should be moved to ListLogins test.

This task was added to this release actually because of the next task
would add one more test file to safe, and there was no letter left to it
(safe already achieved z file test)