site: remove work-a-round from lines ajax
To avoid reload the entire line, a new option of loading just one more
item in the screen was created. But this lost meaning when you need to
put more and more little things to update - e.g. cancellation that will
be implemented in a commit that is coming.

So the line row was refactor in a way that it is entirely replaced in
case something changes in the row.