ci: add comment of why to use different branch
I created a script to add branch and merge it into version branch
because trying to push directly to version branch cause an error in ci
workflow.

But then I forgot why and tried to replace it by a simple push.

To avoid forgetting again, now I added an explanation to the script.