core: remove pack
When I decide to really make satelite solutions get other project via
pack I will need something that will run on satelite solutions and that
build and pack everything in core solution, not depending on the project
to have a script to pack itself.

Right now it is only making the build take longer.